pyproject.toml
[project]
name = "botman"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.11.10",
    "ruff>=0.13.3",
    "python-dotenv>=1.0.0",
    "python-fasthtml>=0.12.29",
    "monsterui>=1.0.29",
]
